release: sync from i4 8ab710e8 (#54)
Sync cosmos-framework with imaginaire4 commit 8ab710e8.

Highlights:
- Promote the new sequence_packing package (modalities, runtime, natten,
packers, types, mrope, temporal_causal) under
cosmos_framework/data/vfm/sequence_packing/, replacing the previous
monolithic sequence_packing.py and the unified_3dmrope_utils helper
(both removed).
- Update consumers (mot/, callbacks/, data/vfm/, omni_mot_model.py) to
the new sequence_packing API.
- Update sft_dataset.py to import add_special_tokens via the new deep
path (cosmos_framework.data.vfm.sequence_packing.modalities) since it is
no longer re-exported from the package root.

---------

+++ cosmos_framework/model/vfm/mot/modeling_utils.py
@@ -367,8 +367,8 @@ class TimestepEmbedder(nn.Module):
         half = dim // 2
         freqs = torch.exp(-math.log(max_period) * torch.arange(start=0, end=half, dtype=torch.float32) / half).to(
             device=t.device
-        )  # [D/2]
-        args = t[:, None].float() * freqs[None]  # [N,D/2]
+        )
+        args = t[:, None] * freqs[None]  # [N,D/2]
         embedding = torch.cat([torch.cos(args), torch.sin(args)], dim=-1)  # [N,D]
         if dim % 2:
             embedding = torch.cat([embedding, torch.zeros_like(embedding[:, :1])], dim=-1)  # [N,D+1]
